TIDIJ-222 Moved map.dart file to map_functionality folder
