+++ fleetcarpooling/lib/pages/map_functionality/udp_message_handler.dart
@@ -12,10 +12,10 @@ class UDPMessageHandler {
       String vehicleContent =
           matches.length > 1 ? matches.elementAt(1).group(1) ?? '' : '';
       String vehicleId = "[${vehicleContent.split('-')[0]}]";
-      
       CustomToast().showStatusToast(message, flagContent, vehicleId);
     } else {
-      CustomToast().showFlutterToast('Uglate zagrade nisu pronađene u tekstu.');
+      CustomToast()
+          .showFlutterToast('Square brackets were not found in the text');
     }
   }
 
@@ -23,9 +23,11 @@ class UDPMessageHandler {
     RegExp regExp = RegExp(r"'latitude': (\d+\.\d+), 'longitude': (\d+\.\d+)");
     var match = regExp.firstMatch(message);
     if (match != null) {
-      var latitude = double.parse(match.group(1)!);
-      var longitude = double.parse(match.group(2)!);
+      final latitude = double.parse(match.group(1)!);
+      final longitude = double.parse(match.group(2)!);
       CustomToast().showFlutterToast('$latitude, $longitude');
+    } else {
+      CustomToast().showFlutterToast('The location format is not correct.');
     }
   }
 }
